% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinding.R
\name{GetDropboxPath}
\alias{GetDropboxPath}
\title{Get Dropbox Path Function}
\usage{
GetDropboxPath(business = F)
}
\arguments{
\item{business}{Business Dropbox or Personal Dropbox? Defaults to FALSE.}
}
\description{
This function allows you to obtain the path for your local dropbox folder.
}
\examples{
GetDropboxPath()
}
\keyword{dropbox}
